requirements.txt
discord.py
openai
python-dotenv
pythonnet
pydub
imageio-ffmpeg
PyNaCl
requests
maturin
ruff